% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comppass.R
\name{comppass}
\alias{comppass}
\title{Calculates Z-Scores, WD scores, and entropy for a given input
table of APMS runs}
\usage{
comppass(input, stats = NULL, norm.factor = 0.98)
}
\arguments{
\item{input}{A table of experiments, replicates, baits, prey, and spectral counts. See details.}

\item{stats}{An optional externally calculated stats table. Must be in the same form as the input table. Defaults to NULL, in which case the stats are calculated from the input table.}

\item{norm.factor}{factor A number between 0 and 1 corresponding to the quantile of the xs to normalize to. Defaults to the 98th percentile.}
}
\value{
A data frame containing the following columns: Experiment.ID, Bait, Prey, AvePSM, Z, WD, Entropy
}
\description{
Calculates Z-Scores, WD scores, and entropy for a given input
table of APMS runs
}
\examples{
data(comppass_test_data)
comppass.test.out <- comppass(comppass.test.data)
}

