% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cRomppass.R
\docType{data}
\name{comppass.test.data}
\alias{comppass.test.data}
\title{Test data for comppass. Includes 27 IP's from the human interactome
project. The column names are as follows:}
\format{A data frame with 27,164 rows and 6 variables}
\usage{
data(comppass_test_data)
}
\description{
\itemize{
  \item Experiment.ID A unique identifier for the AP-MS experiment
  \item Replicate Some label for replicates, eg "A" or "B"
  \item Experiment.Type This is provided for backwards compatability with other software and is ignored
  \item Bait The name of the bait used in the AP-MS experiment
  \item Prey The name of the prey found in the AP-MS experiment
  \item Spectral.Count Number of spectral counts for that given prey in the experiment
}
}
\keyword{datasets}

